Rename chat_scroll.ts to chat_view_reducer.ts
