+++ src/renderer/stores/chat/chat_view_reducer.ts
